mc-9262 refactor tests, making them less DRY, but dealing with an intermittent session context related bug
